jacobian class
Created Jacobian class in its own file, moving existing classes from backend.
Added tests for Finite Difference and Automatic Differentiation Jacobians.
Other Jacobian classes that are unused are unmodified and untested.
